renaming hmm to hidden_markov_model
