+++ dynamax/hidden_markov_model/models/autoregressive_hmm.py
@@ -2,11 +2,9 @@ import jax.numpy as jnp
 import jax.random as jr
 from jax import lax, tree_map
 from typing import NamedTuple
-from dynamax.hmm.models.abstractions import HMM
-from dynamax.hmm.models.initial import StandardHMMInitialState, ParamsStandardHMMInitialState
-from dynamax.hmm.models.transitions import StandardHMMTransitions, ParamsStandardHMMTransitions
-from dynamax.hmm.models.linreg_hmm import (LinearRegressionHMMEmissions,
-                                           ParamsLinearRegressionHMMEmissions)
+from dynamax.hidden_markov_model.models.initial import StandardHMMInitialState, ParamsStandardHMMInitialState
+from dynamax.hidden_markov_model.models.transitions import StandardHMMTransitions, ParamsStandardHMMTransitions
+from dynamax.hidden_markov_model.models.linreg_hmm import LinearRegressionHMMEmissions, ParamsLinearRegressionHMMEmissions
 from dynamax.parameters import ParameterProperties
 from dynamax.utils.bijectors import RealToPSDBijector
